Replace AddressInfo by EthHashInfo
